fixed toggle icon bug
